+++ packages/camera/camera_platform_interface/lib/src/platform_interface/camera_platform.dart
@@ -48,8 +48,8 @@ abstract class CameraPlatform extends PlatformInterface {
 
   /// Creates an uninitialized camera instance and returns the cameraId.
   Future<int> createCamera(
-    CameraDescription cameraDescription,
-    ResolutionPreset? resolutionPreset, {
+    CameraDescription cameraDescription, {
+    MediaSettings? mediaSettings,
     bool enableAudio = false,
   }) {
     throw UnimplementedError('createCamera() is not implemented.');