+++ packages/camera/camera_platform_interface/test/camera_platform_interface_test.dart
@@ -163,7 +163,7 @@ void main() {
             lensDirection: CameraLensDirection.back,
             sensorOrientation: 0,
           ),
-          ResolutionPreset.high,
+          mediaSettings: MediaSettings.low(),
         ),
         throwsUnimplementedError,
       );